164796891-feature: users can create articles
- refactor tests

[Delivers #164796891]
