+++ tests/mock/article.js
@@ -1,0 +1,10 @@
+const ARTICLE = {
+  title: 'Andela is cool 7888',
+  userId: '979eaa2e-5b8f-4103-8192-4639afae2ba8',
+  description: 'Lorem ipsum dolor sit amet, sit ut dolor alterum, sed malis referrentur cu. Aperiam fabulas eos ea. Sea mazim senserit tincidunt te.',
+  body: 'Lorem ipsum dolor sit amet, sit ut dolor alterum, sed malis referrentur cu. Aperiam fabulas eos ea. Sea mazim senserit tincidunt te. Mei volutpat delicatissimi ut, id mollis alienum argumentum has, semper efficiendi sed ea. Ius decore consul forensibus ne, enim verear corpora sit ut. Usu eu possit equidem menandri, quo et noster officiis iracundia.',
+  imageUrl: 'https://picsum.photos/200/300',
+  tags: ['hello', 'async', 'await']
+};
+
+export default ARTICLE;
